moved calls for crypto api to seperate file.
api.rb will be used for all api calls
